feat: better processing and abstraction
