+++ src/transaction.rs
@@ -54,117 +54,6 @@ impl CoreLaneAddresses {
     }
 }
 
-/// Parse Core Lane transaction from Bitcoin DA data
-pub fn parse_core_lane_transaction(data: &[u8]) -> Result<TxEnvelope> {
-    if data.len() < 8 || !data.starts_with(b"CORE_LANE") {
-        return Err(anyhow!(
-            "Invalid Core Lane transaction format - missing CORE_LANE prefix"
-        ));
-    }
-
-    // Extract the Ethereum transaction data (skip CORE_LANE prefix)
-    let tx_data = &data[8..];
-
-    // Try to decode as RLP-encoded Ethereum transaction envelope
-    match TxEnvelope::decode(&mut &tx_data[..]) {
-        Ok(tx) => {
-            // Validate that this is a properly signed transaction
-            match &tx {
-                TxEnvelope::Legacy(signed_tx) => {
-                    // Check if the transaction has a valid signature
-                    if signed_tx.signature().r() == U256::ZERO
-                        && signed_tx.signature().s() == U256::ZERO
-                    {
-                        return Err(anyhow!("Invalid signature: r and s cannot be zero"));
-                    }
-                }
-                TxEnvelope::Eip1559(signed_tx) => {
-                    if signed_tx.signature().r() == U256::ZERO
-                        && signed_tx.signature().s() == U256::ZERO
-                    {
-                        return Err(anyhow!("Invalid signature: r and s cannot be zero"));
-                    }
-                }
-                TxEnvelope::Eip2930(signed_tx) => {
-                    if signed_tx.signature().r() == U256::ZERO
-                        && signed_tx.signature().s() == U256::ZERO
-                    {
-                        return Err(anyhow!("Invalid signature: r and s cannot be zero"));
-                    }
-                }
-                _ => {
-                    // For other transaction types, we'll accept them for now
-                }
-            }
-            Ok(tx)
-        }
-        Err(e) => Err(anyhow!(
-            "Failed to decode Ethereum transaction after CORE_LANE prefix: {}",
-            e
-        )),
-    }
-}
-/// Validate Core Lane transaction
-pub fn validate_transaction(tx: &TxEnvelope) -> Result<()> {
-    // Basic validation - check that we have a valid transaction envelope
-    // In a full implementation, this would validate the signed transaction fields
-    match tx {
-        TxEnvelope::Legacy(_) => debug!("Validating Legacy transaction"),
-        TxEnvelope::Eip2930(_) => debug!("Validating EIP-2930 transaction"),
-        TxEnvelope::Eip1559(_) => debug!("Validating EIP-1559 transaction"),
-        TxEnvelope::Eip4844(_) => debug!("Validating EIP-4844 transaction"),
-        _ => debug!("Validating other transaction type"),
-    }
-
-    Ok(())
-}
-
-/// Calculate transaction gas cost
-pub fn calculate_gas_cost(tx: &TxEnvelope) -> U256 {
-    // For now, return a basic gas cost
-    // In a full implementation, this would calculate actual gas used from the signed transaction
-    match tx {
-        TxEnvelope::Legacy(_) => U256::from(21000u64) * U256::from(1000000000u64), // 21k gas * 1 gwei
-        TxEnvelope::Eip2930(_) => U256::from(21000u64) * U256::from(1000000000u64),
-        TxEnvelope::Eip1559(_) => U256::from(21000u64) * U256::from(1000000000u64),
-        TxEnvelope::Eip4844(_) => U256::from(21000u64) * U256::from(1000000000u64),
-        _ => U256::from(21000u64) * U256::from(1000000000u64),
-    }
-}
-
-/// Extract sender address from transaction signature using alloy's built-in recovery
-pub fn recover_sender(tx: &TxEnvelope) -> Result<Address> {
-    match tx {
-        TxEnvelope::Legacy(signed_tx) => {
-            // Use alloy's built-in signature recovery method
-            match signed_tx.recover_signer() {
-                Ok(address) => Ok(address),
-                Err(e) => Err(anyhow!("Failed to recover signer from Legacy tx: {:?}", e)),
-            }
-        }
-        TxEnvelope::Eip1559(signed_tx) => {
-            // Use alloy's built-in signature recovery method
-            match signed_tx.recover_signer() {
-                Ok(address) => Ok(address),
-                Err(e) => Err(anyhow!(
-                    "Failed to recover signer from EIP-1559 tx: {:?}",
-                    e
-                )),
-            }
-        }
-        TxEnvelope::Eip2930(signed_tx) => {
-            // Use alloy's built-in signature recovery method
-            match signed_tx.recover_signer() {
-                Ok(address) => Ok(address),
-                Err(e) => Err(anyhow!(
-                    "Failed to recover signer from EIP-2930 tx: {:?}",
-                    e
-                )),
-            }
-        }
-        _ => Err(anyhow!("Unsupported transaction type for sender recovery")),
-    }
-}
 /// Transaction execution result
 #[derive(Debug, Clone)]
 pub struct ExecutionResult {
@@ -184,30 +73,6 @@ pub fn execute_transaction(
     //intents: &mut HashMap<B256, (Bytes, u64)>,
     state: &mut CoreLaneState,
 ) -> Result<ExecutionResult> {
-    // Basic execution framework
-    let gas_limit = get_gas_limit(tx);
-    let gas_price = get_gas_price(tx);
-
-    info!("🔄 Executing transaction:");
-    info!("   Sender: {}", sender);
-    info!("   Gas limit: {}", gas_limit);
-    info!("   Gas price: {}", gas_price);
-
-    // Check if sender has enough balance for gas
-    let sender_balance = state.account_manager.get_balance(sender);
-    let max_gas_cost = U256::from(gas_limit) * U256::from(gas_price);
-
-    if sender_balance < max_gas_cost {
-        return Ok(ExecutionResult {
-            success: false,
-            gas_used: U256::ZERO,
-            gas_refund: U256::ZERO,
-            output: Bytes::new(),
-            logs: vec!["Insufficient balance for gas".to_string()],
-            error: Some("Insufficient balance for gas".to_string()),
-        });
-    }
-
     execute_transfer(tx, sender, state)
 }
